prototyping - making it testable
