+++ src/main/java/de/metas/ui/web/bankstatement_reconciliation/actions/ReconcilePaymentsRequest.java
@@ -1,0 +1,48 @@
+package de.metas.ui.web.bankstatement_reconciliation.actions;
+
+import java.util.List;
+
+import javax.annotation.Nullable;
+
+import com.google.common.collect.ImmutableList;
+
+import de.metas.ui.web.bankstatement_reconciliation.BankStatementLineRow;
+import de.metas.ui.web.bankstatement_reconciliation.PaymentToReconcileRow;
+import lombok.Builder;
+import lombok.Builder.Default;
+import lombok.NonNull;
+import lombok.Value;
+
+/*
+ * #%L
+ * metasfresh-webui-api
+ * %%
+ * Copyright (C) 2020 metas GmbH
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 2 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public
+ * License along with this program. If not, see
+ * <http://www.gnu.org/licenses/gpl-2.0.html>.
+ * #L%
+ */
+
+@Value
+@Builder
+public class ReconcilePaymentsRequest
+{
+	@Nullable
+	BankStatementLineRow selectedBankStatementLine;
+
+	@NonNull
+	@Default
+	List<PaymentToReconcileRow> selectedPaymentsToReconcile = ImmutableList.of();
+}
